no compiler warnings

+++ src/dec_mastersetpack.cpp
@@ -148,7 +148,7 @@ static DEC_DECL_PROPAGATESEEED(propagateSeeedMastersetpack)
 
   seeed->setDetectorPropagated(seeedPropagationData->seeedpool->getIndexForDetector(detector));
 
-   if(!seeed->areOpenVarsAndConssCalculated())
+   if( !seeed->areOpenVarsAndConssCalculated() )
    {
       seeed->calcOpenconss();
       seeed->calcOpenvars();
@@ -156,7 +156,7 @@ static DEC_DECL_PROPAGATESEEED(propagateSeeedMastersetpack)
    }
 
    /** set open setpacking constraints to Master */
-   for( size_t i = 0; i < seeed->getNOpenconss(); ++i)
+   for( int i = 0; i < seeed->getNOpenconss(); ++i )
       {
          cons = seeedPropagationData->seeedpool->getConsForIndex(seeed->getOpenconss()[i]);
          if( GCGconsGetType   (cons) == setcovering || GCGconsGetType   (cons) == logicor )
